Choon 11 feature users and login

+++ src/main/java/com/qa/choonz/service/ArtistService.java
@@ -1,15 +1,16 @@
 package com.qa.choonz.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
+import com.qa.choonz.exception.ArtistNotFoundException;
+import com.qa.choonz.persistence.domain.Artist;
+import com.qa.choonz.persistence.domain.UserDetails;
+import com.qa.choonz.persistence.repository.ArtistRepository;
+import com.qa.choonz.persistence.roles.UserRole;
+import com.qa.choonz.rest.dto.ArtistDTO;
 import com.qa.choonz.rest.mapper.ArtistMapper;
 import org.springframework.stereotype.Service;
 
-import com.qa.choonz.exception.ArtistNotFoundException;
-import com.qa.choonz.persistence.domain.Artist;
-import com.qa.choonz.persistence.repository.ArtistRepository;
-import com.qa.choonz.rest.dto.ArtistDTO;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ArtistService {
@@ -23,7 +24,10 @@ public class ArtistService {
         this.mapper = mapper;
     }
 
-    public ArtistDTO create(Artist artist) {
+    public ArtistDTO create(Artist artist, UserDetails user) {
+        // ensure only admin can create artists
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
+
         Artist created = this.repo.save(artist);
         return mapper.mapToDeepDTO(created);
     }
@@ -37,14 +41,20 @@ public class ArtistService {
         return mapper.mapToDeepDTO(found);
     }
 
-    public ArtistDTO update(Artist artist, long id) {
+    public ArtistDTO update(Artist artist, long id, UserDetails user) {
+        // ensure only admin can edit artists
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
+
         Artist toUpdate = this.repo.findById(id).orElseThrow(ArtistNotFoundException::new);
         toUpdate.setName(artist.getName());
         Artist updated = this.repo.save(toUpdate);
         return mapper.mapToDeepDTO(updated);
     }
 
-    public boolean delete(long id) {
+    public boolean delete(long id, UserDetails user) {
+        // ensure only admin can delete artists
+        if (user == null || user.getRole() != UserRole.ADMIN){return false;}
+
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
