+++ src/main/java/com/qa/choonz/service/GenreService.java
@@ -1,15 +1,16 @@
 package com.qa.choonz.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
+import com.qa.choonz.exception.GenreNotFoundException;
+import com.qa.choonz.persistence.domain.Genre;
+import com.qa.choonz.persistence.domain.UserDetails;
+import com.qa.choonz.persistence.repository.GenreRepository;
+import com.qa.choonz.persistence.roles.UserRole;
+import com.qa.choonz.rest.dto.GenreDTO;
 import com.qa.choonz.rest.mapper.GenreMapper;
 import org.springframework.stereotype.Service;
 
-import com.qa.choonz.exception.GenreNotFoundException;
-import com.qa.choonz.persistence.domain.Genre;
-import com.qa.choonz.persistence.repository.GenreRepository;
-import com.qa.choonz.rest.dto.GenreDTO;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class GenreService {
@@ -23,7 +24,10 @@ public class GenreService {
         this.mapper = mapper;
     }
 
-    public GenreDTO create(Genre genre) {
+    public GenreDTO create(Genre genre, UserDetails user) {
+        // ensure only admin can create genre
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
+
         Genre created = this.repo.save(genre);
         return mapper.mapToDeepDTO(created);
     }
@@ -37,7 +41,10 @@ public class GenreService {
         return mapper.mapToDeepDTO(found);
     }
 
-    public GenreDTO update(Genre genre, long id) {
+    public GenreDTO update(Genre genre, long id, UserDetails user) {
+        // ensure only admin can edit genre
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
+
         Genre toUpdate = this.repo.findById(id).orElseThrow(GenreNotFoundException::new);
         toUpdate.setName(genre.getName());
         toUpdate.setDescription(genre.getDescription());
@@ -45,7 +52,10 @@ public class GenreService {
         return mapper.mapToDeepDTO(updated);
     }
 
-    public boolean delete(long id) {
+    public boolean delete(long id, UserDetails user) {
+        // ensure only admin can delete genre
+        if (user == null || user.getRole() != UserRole.ADMIN){return false;}
+
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
